--- code/dimsim/dimsim/__init__.py
@@ -1,1 +1,0 @@
-from .dimsim import *
