+++ dimsim/core/dimsim.py
@@ -1,0 +1,152 @@
+from pypinyin import pinyin, lazy_pinyin, Style
+import sys
+import math
+import itertools
+import os
+
+from dimsim.utils.pinyin import Pinyin, load_pinying_to_simplified, load_pinying_to_traditional
+from dimsim.utils.utils import getEditDistanceClose_TwoDCode, to_pinyin
+from dimsim.utils.maps import vowelMap, consonantMap
+
+doubleConsonantsMap = {}
+doubleVowelsMap = {}
+
+pinyin_to_simplified = load_pinying_to_simplified()
+pinyin_to_traditional = load_pinying_to_traditional()
+
+def get_distance(utterance1, utterance2):
+    if(len(utterance1) is not len(utterance2)):
+        print("the two inputs do not have the same length")
+        return sys.float_info.max
+    else:
+        u1 = to_pinyin(utterance1)
+        u2 = to_pinyin(utterance2)
+        
+        la = []
+        lb = []
+        for py in u1:
+            la.append(Pinyin(py))
+        for py in u2:
+            lb.append(Pinyin(py))
+    
+
+        res = 0.0
+        numDiff = 0        
+        tot = len(utterance1)*2.1
+        for i in range (len(utterance1)):
+            apy = la[i]
+            bpy = lb[i]
+
+            if (apy is None) or (bpy is None):
+                print("!Error {},{}".format(la, lb))
+            
+            res += getEditDistanceClose_TwoDCode(apy, bpy)
+            
+            if apy.consonant is not bpy.consonant:
+                numDiff+=1
+            
+            if not(str(apy.vowel) == str(bpy.vowel)):
+                numDiff+=1
+            
+            if apy.tone is not bpy.tone:
+                numDiff+=0.01
+                
+        diffRatio = (numDiff)/tot
+        a = 0
+        if diffRatio is 0:
+            a=1
+        return res*diffRatio      
+        
+
+def getCandidates(sentence, mode="simplified", theta=1):
+    candidates = []
+    words_candidates = []
+    for word in sentence:
+        candid = __getClosePinyinCandids(word, theta)
+        words_candidates.append(candid)
+    all_combinations = itertools.product(*words_candidates)
+    counter = 0
+    for combination in all_combinations:
+        counter+=1
+        searchKey = ""
+        for i in combination:
+            searchKey = searchKey + i.toStringWithTone().replace("None","") + " "
+        if mode is "simplified":
+            if searchKey.strip() in pinyin_to_simplified:
+                candidates+=pinyin_to_simplified[searchKey.strip()]
+        else:
+            if searchKey.strip() in pinyin_to_traditional:
+                candidates+=pinyin_to_traditional[searchKey.strip()]
+    return candidates
+
+
+def __getClosePinyinCandids(word, theta=2):
+    res = []
+    word_pinyin = to_pinyin(word)
+    word_py = Pinyin(word_pinyin[0])
+    
+    cCandids = __getConsonantCandids(theta, word_py)
+    for i in range(len(cCandids)):
+        if cCandids[i] == word_py.consonant:
+            continue
+        for j in range(1,5,1):
+            newPy = cCandids[i]+word_py.vowel+str(j)
+            res.append(Pinyin(newPy))
+    
+    vCandids = __getVowelCandids(theta, word_py)
+    for i in range(len(vCandids)):
+        for j in range(1,5,1):
+            if word_py.consonant is None:
+                newPy = vCandids[i]+str(j)
+            else:
+                newPy = word_py.consonant+vCandids[i]+str(j)
+            res.append(Pinyin(newPy))
+    return res
+    
+def __getConsonantCandids(theta, word_py):
+    __populateDoubleConsonantsMap()
+    res = []
+    curCode = 0        
+    if word_py.consonant is None:
+        orgCode = consonantMap["__v"]
+    else:
+        orgCode = consonantMap[word_py.consonant]
+        for i in range(int(orgCode-theta), int(orgCode+theta), 1):
+            if float(i) in doubleConsonantsMap:
+                cand = doubleConsonantsMap[float(i)]
+                if cand is not None:
+                    res += cand
+    return res
+    
+
+def __getVowelCandids(theta, word_py):
+    __populateDoubleVowelsMap()
+    res = []
+    curCode = 0        
+    orgCode = vowelMap[word_py.vowel]
+    for i in range(int(orgCode-theta), int(orgCode+theta), 1):
+        if float(i) in doubleVowelsMap:
+            cand = doubleVowelsMap[float(i)]
+            if cand is not None:
+                res += cand
+    return res
+
+def __populateDoubleConsonantsMap():
+    if len(doubleConsonantsMap) is not 0:
+        return
+    hmCdouble = consonantMap
+    for consonant in hmCdouble:
+        if hmCdouble[consonant] not in doubleConsonantsMap:
+            doubleConsonantsMap[hmCdouble[consonant]] = []
+            
+        doubleConsonantsMap[hmCdouble[consonant]].append(consonant)
+        
+def __populateDoubleVowelsMap():
+    if len(doubleVowelsMap) is not 0:
+        return
+    hmVdouble = vowelMap
+    for vowel in hmVdouble:
+        if hmVdouble[vowel] not in doubleVowelsMap:
+            doubleVowelsMap[hmVdouble[vowel]] = []
+            
+        doubleVowelsMap[hmVdouble[vowel]].append(vowel)
