+++ dimsim/data/pinyin_to_simplified.pickle
