+++ dimsim/data/pinyin_to_traditional.pickle
